LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-in-foreign-domain-subframe-location-change-expected.txt

CONSOLE MESSAGE: line 41: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
Opener frame


PASS: Cross frame access to a data: URL embed in a frame on a foreign domain denied!


--------
Frame: '<!--frame1-->'
--------
PASS: Cross frame access from a frame on a foreign domain denied!

Inner iframe. This iframe (which is data: URL and whose parent is on a foreign domain) is the frame that the main frame is trying to access. It should not have access to it.
